+++ Classes/ARM/Restore/RelativityArmRestoreJobCreateResponse.ps1
@@ -1,0 +1,10 @@
+class RelativityArmRestoreJobCreateResponse 
+{
+    [ValidateNotNull()]
+    [Int32] $JobID
+
+    RelativityArmRestoreJobCreateResponse([Int32] $jobId)
+    {
+        $this.JobID = $jobID
+    }
+}