+++ Private/RelativityAPI/Get-RelativityApiRequestHeader.ps1
@@ -1,35 +1,5 @@
 <#
 .SYNOPSIS
-Gets the base endpoint for a specified Relativity business domain.
-
-.DESCRIPTION
-Constructs and returns the base API endpoint URL for the provided business domain in the Relativity system.
-
-.PARAMETER RelativityBusinessDomain
-The business domain for which the base endpoint URL is to be constructed. Currently, only "ARM" is supported.
-
-.NOTES
-This function is internal and is used to construct full API endpoints for other operations.
-#>
-function Get-RelativityApiEndpointBase
-{
-    Param
-    (
-        [Parameter(Mandatory = $true)]
-        [ValidateSet("ARM")]
-        [String] $RelativityBusinessDomain
-    )
-
-    switch ($RelativityBusinessDomain)
-    {
-        "ARM" { $RelativityBusinessDomainEndpoint = "relativity-arm/v1" }
-    }
-
-    return "$($RelativityBaseUri)/Relativity.REST/api/$($RelativityBusinessDomainEndpoint)/"
-}
-
-<#
-.SYNOPSIS
 Generates the headers required for Relativity API requests.
 
 .DESCRIPTION