+++ Private/RelativityAPI/Invoke-RelativityApiRequest.ps1
@@ -5,11 +5,8 @@ Makes an API request to the specified Relativity endpoint.
 .DESCRIPTION
 Constructs and sends an API request to the given Relativity endpoint, using the specified HTTP method and request body.
 
-.PARAMETER RelativityBusinessDomain
-The business domain for which the API request is intended.
-
-.PARAMETER RelativityApiEndpointExtended
-The specific API endpoint (excluding the base) to which the request is to be made.
+.PARAMETER RelativityApiEndpoint
+The specific API endpoint to which the request is to be made.
 
 .PARAMETER RelativityApiHttpMethod
 The HTTP method to use for the API request. Currently, only "Post" is supported.
@@ -25,14 +22,12 @@ function Invoke-RelativityApiRequest
     Param
     (
         [Parameter(Mandatory = $true)]
-        [ValidateSet("ARM")]
-        [String] $RelativityBusinessDomain,
+        [ValidateNotNullOrEmpty()]
+        [String] $RelativityApiEndpoint,
         [Parameter(Mandatory = $true)]
-        [String] $RelativityApiEndpointExtended,
-        [Parameter(Mandatory = $true)]
-        [ValidateSet("Get", "Post")]
+        [ValidateSet("Post", "Get", "Put", "Delete")]
         [String] $RelativityApiHttpMethod,
-        [Parameter(Mandatory = $true)]
+        [Parameter(Mandatory = $false)]
         [Hashtable] $RelativityApiRequestBody
     )
 
@@ -46,17 +41,16 @@ function Invoke-RelativityApiRequest
         throw "RelativityCredential is not set. Please run Set-RelativityCredential before proceeding."
     }
 
-    $RelativityApiEndpointBase = Get-RelativityApiEndpointBase -RelativityBusinessDomain $RelativityBusinessDomain
-    $RelativityApiEndpoint = "$($RelativityApiEndpointBase)$($RelativityApiEndpointExtended)"
-
     $RelativityApiRequestHeader = Get-RelativityApiRequestHeader
 
     try
     {
         switch ($RelativityApiHttpMethod)
         {
-            "Get" { $RelativityApiResponse = Invoke-RestMethod -Uri $RelativityApiEndpoint -Method Get -Headers $RelativityApiRequestHeader }
             "Post" { $RelativityApiResponse = Invoke-RestMethod -Uri $RelativityApiEndpoint -Method Post -Headers $RelativityApiRequestHeader -Body ($RelativityApiRequestBody | ConvertTo-Json -Depth 3) -ContentType "application/json" }
+            "Get" { $RelativityApiResponse = Invoke-RestMethod -Uri $RelativityApiEndpoint -Method Get -Headers $RelativityApiRequestHeader }
+            "Put" { $RelativityApiResponse = Invoke-RestMethod -Uri $RelativityApiEndpoint -Method Put -Headers $RelativityApiRequestHeader -Body ($RelativityApiRequestBody | ConvertTo-Json -Depth 3) -ContentType "application/json" }
+            "Delete" { $RelativityApiResponse = Invoke-RestMethod -Uri $RelativityApiEndpoint -Method Delete -Headers $RelativityApiRequestHeader }
         }
     }
     catch